+++ src/notcurses/methods.rs
@@ -600,7 +600,7 @@ impl Nc {
         error![
             unsafe { c_api::notcurses_refresh(self, &mut y, &mut x) },
             "",
-            (y as u32, x as u32)
+            (y, x)
         ]
     }
 
@@ -828,12 +828,12 @@ impl Nc {
         // if an `NcVisual` is not provided, only `maxpixel_yx`, `cdim_yx` and
         // `scale_yx` can be non-zero.
         if visual.is_none() {
-            cdim_yx = Some((vg.cdimy as u32, vg.cdimx as u32));
-            scale_yx = Some((vg.scaley as u32, vg.scalex as u32));
+            cdim_yx = Some((vg.cdimy, vg.cdimx));
+            scale_yx = Some((vg.scaley, vg.scalex));
 
             // pixel blitter only is defined for Ncblitter::PIXEL
             if vg.blitter == NcBlitter::Pixel.into() {
-                maxpixel_yx = Some((vg.maxpixely as u32, vg.maxpixelx as u32));
+                maxpixel_yx = Some((vg.maxpixely, vg.maxpixelx));
             } else {
                 maxpixel_yx = None;
             }
@@ -846,7 +846,7 @@ impl Nc {
         } else {
             // `maxpixel_yx` only is defined for `Ncblitter`::PIXEL.
             if vg.blitter == NcBlitter::Pixel.into() {
-                maxpixel_yx = Some((vg.maxpixely as u32, vg.maxpixelx as u32));
+                maxpixel_yx = Some((vg.maxpixely, vg.maxpixelx));
             } else {
                 maxpixel_yx = None;
             }
@@ -856,35 +856,35 @@ impl Nc {
                 beg_yx = None;
                 len_yx = None;
             } else {
-                beg_yx = Some((vg.begy as u32, vg.begx as u32));
-                len_yx = Some((vg.leny as u32, vg.lenx as u32));
+                beg_yx = Some((vg.begy, vg.begx));
+                len_yx = Some((vg.leny, vg.lenx));
             }
 
             // valid values for the following fields can't be 0 either:
             if vg.pixy | vg.pixx == 0 {
                 pix_yx = None;
             } else {
-                pix_yx = Some((vg.pixy as u32, vg.pixx as u32));
+                pix_yx = Some((vg.pixy, vg.pixx));
             }
             if vg.cdimy | vg.cdimx == 0 {
                 cdim_yx = None;
             } else {
-                cdim_yx = Some((vg.cdimy as u32, vg.cdimx as u32));
+                cdim_yx = Some((vg.cdimy, vg.cdimx));
             }
             if vg.scaley | vg.scalex == 0 {
                 scale_yx = None;
             } else {
-                scale_yx = Some((vg.scaley as u32, vg.scalex as u32));
+                scale_yx = Some((vg.scaley, vg.scalex));
             }
             if vg.rpixy | vg.rpixx == 0 {
                 rpix_yx = None;
             } else {
-                rpix_yx = Some((vg.rpixy as u32, vg.rpixx as u32));
+                rpix_yx = Some((vg.rpixy, vg.rpixx));
             }
             if vg.rcelly | vg.rcellx == 0 {
                 rcell_yx = None;
             } else {
-                rcell_yx = Some((vg.rcelly as u32, vg.rcellx as u32));
+                rcell_yx = Some((vg.rcelly, vg.rcellx));
             }
         }
